Fixed some issues with transformer not returning in correct format
